refactor(env): revert back en
