+++ packages/ui/lib/components/index.ts
@@ -1,2 +1,1 @@
 export * from './ui';
-export * from './smart';
