--- packages/ui/lib/components/smart/index.ts
@@ -1,1 +1,0 @@
-export * from './api-error-handler';
