+++ pages/popup/src/Popup.tsx
@@ -2,22 +2,22 @@ import '@src/Popup.css';
 
 import { withErrorBoundary, withSuspense } from '@extension/shared';
 import { store, ReduxProvider } from '@extension/store';
-import { ApiErrorHandler } from '@extension/ui';
 
-import { AuthGuard } from './components/guards';
 import { Skeleton } from './components/ui';
+import { AuthGuard } from './guards';
 import { PopupContent } from './popup-content';
+import { ApiHealthProvider } from './providers';
 
 const Popup = () => (
-  <ReduxProvider store={store}>
-    <div className="dark:bg-background relative px-5 pb-5 pt-4">
-      <ApiErrorHandler>
+  <ApiHealthProvider>
+    <ReduxProvider store={store}>
+      <div className="dark:bg-background relative px-5 pb-5 pt-4">
         <AuthGuard>
           <PopupContent />
         </AuthGuard>
-      </ApiErrorHandler>
-    </div>
-  </ReduxProvider>
+      </div>
+    </ReduxProvider>
+  </ApiHealthProvider>
 );
 
 export default withErrorBoundary(withSuspense(Popup, <Skeleton />), <div>Error Occurred</div>);
