+++ pages/popup/src/guards/auth.guard.tsx
@@ -3,7 +3,7 @@ import type { PropsWithChildren } from 'react';
 import { useStorage } from '@extension/shared';
 import { authTokensStorage } from '@extension/storage';
 
-import { AuthView } from '../ui';
+import { AuthView } from '../components/ui';
 
 export const AuthGuard: React.FC<PropsWithChildren> = ({ children }) => {
   const tokens = useStorage(authTokensStorage);
