+++ pages/popup/src/guards/index.ts
