+++ pages/popup/src/providers/index.ts
@@ -1,0 +1,1 @@
+export { ApiHealthProvider } from './api-health.provider';
